Fixed make strict build and cleaned up some warnings

+++ src/protocol/inspircd11.c
@@ -900,7 +900,7 @@ void inspircd_cmd_pong(char *servname, char *who)
 /* JOIN */
 void inspircd_cmd_join(char *user, char *channel, time_t chantime)
 {
-    send_cmd(user, "JOIN %s %ld", channel, (unsigned int)chantime);
+    send_cmd(user, "JOIN %s %ld", channel, chantime);
 }
 
 /* UNSQLINE */
